requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
